Update file creation messages

+++ src/js/modules/commands/install/create/files/composer.js
@@ -5,6 +5,7 @@
 
 // Import node modules
 import path from "node:path";
+import { existsSync } from "node:fs";
 import fs from "node:fs/promises";
 
 // Import helpers
@@ -16,8 +17,6 @@ import { c, log, rootDirPath, runCommand } from "../../../../utils/helpers/index
  * @param {string} wpCoreDir The dirname for the WordPress Core directory.
  */
 export async function createComposerFile(wpCoreDir) {
-  log(c.detail("Creating `composer.json` file..."));
-
   const composerTarget = path.join(rootDirPath, "composer.json");
 
   const composerFileContent = `{
@@ -52,7 +51,7 @@ export async function createComposerFile(wpCoreDir) {
 }`;
 
   const updateComposerWPInstallDir = async () => {
-    log(c.detail("Updating WordPress Core directory path in composer.json..."));
+    // log(c.detail("Updating WordPress Core directory path in composer.json..."));
     wpCoreDir = wpCoreDir.replace(/^\//, "");
     const installFilePath = path.join(rootDirPath, "composer.json");
     const file = await fs.readFile(installFilePath, "utf8");
@@ -66,16 +65,14 @@ export async function createComposerFile(wpCoreDir) {
     await fs.writeFile(installFilePath, JSON.stringify(composerFile, null, 2) + "\n", "utf8");
   };
 
-  try {
-    await fs.access(composerTarget);
-    // composer.json file already exists...
+  if (existsSync(composerTarget)) {
     log(
-      c.info(
-        "The composer.json file already exists. Skipping creation and adding the `roots/wordpress` composer package to the existing composer.json file...",
+      c.dim(
+        `The ${c.em(`${"composer.json"}`)} file already exists. Skipping creation and adding the \`roots/wordpress\` composer package to the existing composer.json file.\n`,
       ),
     );
-  } catch {
-    // composer.json file does not exist...
+  } else {
+    log(c.detail(`Creating ${c.em("composer.json")} file...`));
     await fs.writeFile(composerTarget, composerFileContent, "utf8");
   }
